+++ Source/SwiftLintFramework/StyleViolationType.swift
@@ -7,17 +7,18 @@
 //
 
 public enum StyleViolationType: String, Printable {
-    case NameFormat            = "Name Format"
-    case Length                = "Length"
-    case TrailingNewline       = "Trailing Newline"
-    case LeadingWhitespace     = "Leading Whitespace"
-    case TrailingWhitespace    = "Trailing Whitespace"
-    case ReturnArrowWhitespace = "Return Arrow Whitespace"
-    case ForceCast             = "Force Cast"
-    case TODO                  = "TODO or FIXME"
-    case Colon                 = "Colon"
-    case Nesting               = "Nesting"
-    case ControlStatement      = "Control Statement Parentheses"
+    case NameFormat                 = "Name Format"
+    case Length                     = "Length"
+    case TrailingNewline            = "Trailing Newline"
+    case LeadingWhitespace          = "Leading Whitespace"
+    case TrailingWhitespace         = "Trailing Whitespace"
+    case ReturnArrowWhitespace      = "Return Arrow Whitespace"
+    case OperatorFunctionWhitespace = "Operator Function Whitespace"
+    case ForceCast                  = "Force Cast"
+    case TODO                       = "TODO or FIXME"
+    case Colon                      = "Colon"
+    case Nesting                    = "Nesting"
+    case ControlStatement           = "Control Statement Parentheses"
 
     public var description: String { return rawValue }
 }