+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 ##### Enhancements
 
+* Added `OperatorFunctionWhitespaceRule` to make sure that
+  you use whitespace around operators when defining them.  
+  [Akira Hirakawa](https://github.com/akirahrkw)
+  [#60](https://github.com/realm/SwiftLint/issues/60)
+
 * Added `ReturnArrowWhitespaceRule` to make sure that
   you have 1 space before return arrow and return type.  
   [Akira Hirakawa](https://github.com/akirahrkw)